rename : front merge 위한 파일 이동
